int to double... turns out pi is not equal to 3

+++ DirectProgramming/DPC++/MapReduce/MonteCarloPi/src/monte_carlo_pi.cpp
@@ -66,7 +66,7 @@ void MonteCarloPi(rgb * image_plot){
     }
 
     // Print calculated value of pi
-    int pi = 4.0 * (double) count / size_n;
+    double pi = 4.0 * (double) count / size_n;
     std::cout << "Computation complete. The estimated value of pi is: " << pi << std::endl;
 }
